PMM-5540 Use simple structures for prometheus config parsing
