--- services/prometheus/internal/prometheus/NOTICE
@@ -1,5 +1,0 @@
-The Prometheus systems and service monitoring server
-Copyright 2012-2015 The Prometheus Authors
-
-This product includes software developed at
-SoundCloud Ltd. (http://soundcloud.com/).
