--- services/prometheus/internal/prometheus/discovery/targetgroup/targetgroup.go
@@ -1,111 +1,0 @@
-// pmm-managed
-// Copyright (C) 2017 Percona LLC
-//
-// This program is free software: you can redistribute it and/or modify
-// it under the terms of the GNU Affero General Public License as published by
-// the Free Software Foundation, either version 3 of the License, or
-// (at your option) any later version.
-//
-// This program is distributed in the hope that it will be useful,
-// but WITHOUT ANY WARRANTY; without even the implied warranty of
-// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
-// GNU Affero General Public License for more details.
-//
-// You should have received a copy of the GNU Affero General Public License
-// along with this program. If not, see <https://www.gnu.org/licenses/>.
-
-// Original file: https://github.com/prometheus/prometheus/blob/v2.16.0/discovery/targetgroup/targetgroup.go
-
-// Copyright 2013 The Prometheus Authors
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-// http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package targetgroup
-
-import (
-	"bytes"
-	"encoding/json"
-
-	"github.com/prometheus/common/model"
-)
-
-// Group is a set of targets with a common label set(production , test, staging etc.).
-type Group struct {
-	// Targets is a list of targets identified by a label set. Each target is
-	// uniquely identifiable in the group by its address label.
-	Targets []model.LabelSet
-	// Labels is a set of labels that is common across all targets in the group.
-	Labels model.LabelSet
-
-	// Source is an identifier that describes a group of targets.
-	Source string
-}
-
-func (tg Group) String() string {
-	return tg.Source
-}
-
-// UnmarshalYAML implements the yaml.Unmarshaler interface.
-func (tg *Group) UnmarshalYAML(unmarshal func(interface{}) error) error {
-	g := struct {
-		Targets []string       `yaml:"targets"`
-		Labels  model.LabelSet `yaml:"labels"`
-	}{}
-	if err := unmarshal(&g); err != nil {
-		return err
-	}
-	tg.Targets = make([]model.LabelSet, 0, len(g.Targets))
-	for _, t := range g.Targets {
-		tg.Targets = append(tg.Targets, model.LabelSet{
-			model.AddressLabel: model.LabelValue(t),
-		})
-	}
-	tg.Labels = g.Labels
-	return nil
-}
-
-// MarshalYAML implements the yaml.Marshaler interface.
-func (tg Group) MarshalYAML() (interface{}, error) {
-	g := &struct {
-		Targets []string       `yaml:"targets"`
-		Labels  model.LabelSet `yaml:"labels,omitempty"`
-	}{
-		Targets: make([]string, 0, len(tg.Targets)),
-		Labels:  tg.Labels,
-	}
-	for _, t := range tg.Targets {
-		g.Targets = append(g.Targets, string(t[model.AddressLabel]))
-	}
-	return g, nil
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (tg *Group) UnmarshalJSON(b []byte) error {
-	g := struct {
-		Targets []string       `json:"targets"`
-		Labels  model.LabelSet `json:"labels"`
-	}{}
-
-	dec := json.NewDecoder(bytes.NewReader(b))
-	dec.DisallowUnknownFields()
-	if err := dec.Decode(&g); err != nil {
-		return err
-	}
-	tg.Targets = make([]model.LabelSet, 0, len(g.Targets))
-	for _, t := range g.Targets {
-		tg.Targets = append(tg.Targets, model.LabelSet{
-			model.AddressLabel: model.LabelValue(t),
-		})
-	}
-	tg.Labels = g.Labels
-	return nil
-}
